requirements.txt
onnxruntime
Pillow
flask
waitress
opencv-python